Added coffee houses list request

+++ app/src/main/java/com/example/hotcoffee/data/remote/models/Location.kt
@@ -1,6 +1,6 @@
 package com.example.hotcoffee.data.remote.models
 
 data class Location(
-    val latitude: Int,
-    val longitude: Int
+    val latitude: Double,
+    val longitude: Double
 )
